third_party/blink/web_tests/platform/generic/http/tests/eventsource/eventsource-parse-event-stream-expected.txt

Test EventSource text/event-stream parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS es = new EventSource("resources/event-stream.php"); did not throw exception.
PASS got open event
PASS evt.data is "\n\n"
PASS evt.data is "simple"
PASS evt.data is "spanning\nmultiple\n\nlines\n"
PASS evt.data is "id is 1"
PASS evt.lastEventId is "1"
PASS evt.data is "id is still 1"
PASS evt.lastEventId is "1"
PASS evt.data is "no id"
PASS evt.lastEventId is ""
PASS got open event from server
PASS evt.data is "a message event with the name \"message\""
PASS evt.data is "a line ending with crlf\na line with a : (colon)\na line ending with cr"
PASS count is 7
DONE
PASS successfullyParsed is true

TEST COMPLETE
